fraud_detector/requirements.txt
catboost==1.2.8
contourpy==1.0.7
cycler==0.12.1
fonttools==4.38.0
geographiclib==2.0
geopy==2.4.1
graphviz==0.20.1
joblib==1.3.2
kiwisolver==1.4.4
matplotlib==3.7.3
numpy==1.25.2
packaging==24.2
pandas==2.1.1
pyparsing==3.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.1
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.12.2
six==1.17.0
threadpoolctl==3.2.0
tzdata==2025.2
watchdog==3.0.0
kafka-python==2.2.6
# psycopg2-binary==2.9.9
